Divide general helpers and client helpers
